+++ app/src/main/java/com/progmobile/meetchup/utils/views/Icon.java
@@ -1,0 +1,29 @@
+package com.progmobile.meetchup.utils.views;
+
+import android.content.Context;
+import android.util.AttributeSet;
+
+import androidx.appcompat.widget.AppCompatImageView;
+
+
+public class Icon extends AppCompatImageView {
+
+    public Icon(final Context context) {
+        super(context);
+    }
+
+    public Icon(final Context context, final AttributeSet attrs) {
+        super(context, attrs);
+    }
+
+    public Icon(final Context context, final AttributeSet attrs, final int defStyle) {
+        super(context, attrs, defStyle);
+    }
+
+    @Override
+    protected void onMeasure(int width, int height) {
+        super.onMeasure(width, height);
+        int measuredWidth = getMeasuredWidth();
+        setMeasuredDimension(measuredWidth, measuredWidth);
+    }
+}
