+++ app/src/main/res/layout/item_photo.xml
@@ -6,7 +6,7 @@
     android:gravity="center"
     >
 
-    <com.progmobile.meetchup.utils.Icon
+    <com.progmobile.meetchup.utils.views.Icon
         android:id="@+id/item_photo_image"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
